// client/src/api/model/talentTier.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type TalentTier = typeof TalentTier[keyof typeof TalentTier];


export const TalentTier = {
  First: 'First',
  Second: 'Second',
  Third: 'Third',
  Fourth: 'Fourth',
  Fifth: 'Fifth',
} as const;